requirements.txt
pillow
PySide2
qrcode
